module04_vector_databases/vector_db_implementations.py
# vector_db_implementations.py
# Explanation of Vector Database Implementations
# Popular vector databases that provide efficient storage and retrieval of embeddings.
# These are production-ready systems that handle millions of vectors with advanced indexing.

# Note: This file demonstrates concepts; actual usage requires installing specific libraries
# All examples are commented out to avoid dependency issues in the learning environment

# Chroma - Open-source embedding database
# Lightweight vector database that runs locally or in containers
def chroma_example():
    """
    Example of using Chroma vector database.
    Chroma is designed for AI applications and provides simple APIs for embedding storage.
    """
    # Try to import Chroma library
    try:
        import chromadb

        # Create a Chroma client (can be persistent or in-memory)
        # Client handles connection to Chroma server or local instance
        client = chromadb.Client()

        # Create a named collection to store vectors and metadata
        # Collections are like tables in traditional databases
        collection = client.create_collection("my_collection")

        # Prepare sample documents and their embeddings
        # In real usage, embeddings would come from models like sentence-transformers
        documents = ["This is a document about cats", "This is about dogs"]
        # Dummy 3D embeddings for demonstration (real embeddings are 384+ dimensions)
        embeddings = [[0.1, 0.2, 0.3], [0.4, 0.5, 0.6]]
        # Unique IDs for each document/vector pair
        ids = ["doc1", "doc2"]

        # Add documents, embeddings, and IDs to the collection
        # Chroma automatically handles indexing for efficient similarity search
        collection.add(
            embeddings=embeddings,    # The vector representations
            documents=documents,      # Original text content
            ids=ids                   # Unique identifiers
        )

        # Create a query embedding (would normally be generated from user query)
        query_embedding = [0.15, 0.25, 0.35]

        # Perform similarity search using the query embedding
        # n_results specifies how many similar documents to return
        results = collection.query(
            query_embeddings=[query_embedding],  # Query vector(s)
            n_results=2                          # Number of results to return
        )

        # Display search results
        print("Chroma Results:")
        # Results contain documents and distances (lower distance = more similar)
        for doc, score in zip(results['documents'][0], results['distances'][0]):
            print(f"Document: {doc}, Distance: {score}")

    # Handle case where Chroma is not installed
    except ImportError:
        print("Chroma not installed. Install with: pip install chromadb")

# Pinecone - Managed vector database
# Cloud-hosted vector database with automatic scaling and management
def pinecone_example():
    """
    Example of using Pinecone (requires API key).
    Pinecone is a fully managed service that handles infrastructure and scaling.
    """
    # Try to import Pinecone client
    try:
        from pinecone import Pinecone

        # Initialize Pinecone client with API key (required for authentication)
        # pc = Pinecone(api_key="your-api-key")

        # Create a new index with specified dimensions and similarity metric
        # dimension: size of vectors (must match your embedding model)
        # metric: similarity measure (cosine, euclidean, dotproduct)
        # pc.create_index("my-index", dimension=128, metric="cosine")

        # Connect to an existing index for operations
        # index = pc.Index("my-index")

        # Insert vectors into the index (upsert = update if exists, insert if not)
        # Each vector needs: ID, vector values, optional metadata
        # index.upsert([
        #     ("vec1", [0.1, 0.2, ...], {"metadata": "value"}),
        #     ("vec2", [0.4, 0.5, ...], {"metadata": "value"})
        # ])

        # Query the index for similar vectors
        # vector: query embedding, top_k: number of results
        # results = index.query(vector=[0.1, 0.2, ...], top_k=3)

        # Print information about Pinecone's features
        print("Pinecone: Managed cloud vector database")
        print("- Serverless architecture - no infrastructure management needed")
        print("- Automatic scaling - handles traffic spikes automatically")
        print("- RESTful API - simple HTTP interface for all operations")

    # Handle case where Pinecone client is not installed
    except ImportError:
        print("Pinecone not installed. Install with: pip install pinecone-client")

# Weaviate - Open-source vector database with GraphQL API
# Feature-rich vector database with advanced querying capabilities
def weaviate_example():
    """
    Example of using Weaviate.
    Weaviate combines vector search with traditional database features.
    """
    # Try to import Weaviate client
    try:
        import weaviate

        # Connect to a Weaviate instance (local or cloud)
        # client = weaviate.Client("http://localhost:8080")

        # Define a schema for the data model
        # Schema specifies classes (like tables) and their properties
        # schema = {
        #     "classes": [{
        #         "class": "Document",
        #         "properties": [
        #             {"name": "content", "dataType": ["string"]},      # Text content
        #             {"name": "embedding", "dataType": ["number[]"]}  # Vector embedding
        #         ]
        #     }]
        # }

        # Create the schema in Weaviate
        # client.schema.create(schema)

        # Add data objects with vector embeddings
        # Objects combine traditional data with vector representations
        # client.data_object.create({
        #     "content": "Document text",
        #     "embedding": [0.1, 0.2, 0.3]
        # }, "Document")

        # Perform vector similarity search using GraphQL-like syntax
        # with_near_vector finds objects similar to the given vector
        # result = client.query.get("Document", ["content"]).with_near_vector({
        #     "vector": [0.1, 0.2, 0.3]
        # }).do()

        # Print information about Weaviate's features
        print("Weaviate: Open-source vector database")
        print("- GraphQL and REST APIs - flexible query interfaces")
        print("- Built-in classification and clustering - advanced ML features")
        print("- Schema-based data modeling - structured data with vectors")

    # Handle case where Weaviate client is not installed
    except ImportError:
        print("Weaviate not installed. Install with: pip install weaviate-client")

# Run all vector database examples
print("Vector Database Implementations:")
chroma_example()   # Demonstrate Chroma usage
print()
pinecone_example() # Demonstrate Pinecone concepts
print()
weaviate_example() # Demonstrate Weaviate concepts

# Comparison of different vector database options:
# - Chroma: Lightweight, easy to use, good for small projects and prototyping
# - Pinecone: Managed service, scalable, good for production applications
# - Weaviate: Feature-rich, supports complex queries and relationships between data